+++ src/classes/job.ts
@@ -543,7 +543,7 @@ export class Job<
     if (
       this.attemptsMade < this.opts.attempts &&
       !this.discarded &&
-      !(err instanceof UnrecoverableError)
+      !(err instanceof UnrecoverableError || err.name == 'UnrecoverableError')
     ) {
       const opts = queue.opts as WorkerOptions;
 
@@ -568,7 +568,9 @@ export class Job<
         command = 'delayed';
       } else {
         // Retry immediately
-        (<any>multi).retryJob(this.scripts.retryJobArgs(this));
+        (<any>multi).retryJob(
+          this.scripts.retryJobArgs(this.id, this.opts.lifo, token),
+        );
         command = 'retry';
       }
     } else {
@@ -715,8 +717,9 @@ export class Job<
       ];
 
       const transformedProcessed = Object.entries(processed).reduce(
-        (accumulator, [key, value]) => {
-          return { ...accumulator, [key]: JSON.parse(value) };
+        (accumulator: Record<string, any>, [key, value]) => {
+          accumulator[key] = JSON.parse(value);
+          return accumulator;
         },
         {},
       );